Remove all unnecessary code as per sonarqube suggestions

+++ music-core/src/main/java/com/sismics/util/FilenameUtil.java
@@ -22,7 +22,7 @@ public class FilenameUtil {
     public static String cleanFileName(String badFileName) {
         StringBuilder cleanName = new StringBuilder();
         for (int i = 0; i < badFileName.length(); i++) {
-            int c = (int) badFileName.charAt(i);
+            int c = badFileName.charAt(i);
             if (Arrays.binarySearch(illegalChars, c) < 0) {
                 cleanName.append((char) c);
             }
